refactor : refactor acceptance test and controller
