--- src/main/java/nextstep/subway/favorite/application/MemberFavoriteService.java
@@ -1,81 +1,0 @@
-package nextstep.subway.favorite.application;
-
-import nextstep.subway.auth.exception.AuthorizationException;
-import nextstep.subway.favorite.domain.Favorite;
-import nextstep.subway.favorite.domain.FavoriteRepository;
-import nextstep.subway.favorite.dto.FavoriteRequest;
-import nextstep.subway.favorite.dto.FavoriteResponse;
-import nextstep.subway.favorite.exception.NotExistException;
-import nextstep.subway.station.domain.Station;
-import nextstep.subway.station.domain.StationRepository;
-import nextstep.subway.station.dto.StationResponse;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.function.Function;
-import java.util.stream.Collectors;
-
-@Service
-@Transactional
-public class MemberFavoriteService {
-    private final FavoriteRepository favoriteRepository;
-    private final StationRepository stationRepository;
-
-    public MemberFavoriteService(FavoriteRepository favoriteRepository, StationRepository stationRepository) {
-        this.favoriteRepository = favoriteRepository;
-        this.stationRepository = stationRepository;
-    }
-
-    public Long createFavorite(Long memberId, FavoriteRequest request) {
-        Favorite favorite = request.toEntity(memberId);
-        Favorite savedFavorite = favoriteRepository.save(favorite);
-        return savedFavorite.getId();
-    }
-
-    public void deleteFavorite(Long memberId, Long favoriteId) {
-        Favorite favorite = favoriteRepository.findById(favoriteId).orElseThrow(NotExistException::new);
-
-        if (!favorite.isCreatedBy(memberId)) {
-            throw new AuthorizationException();
-        }
-
-        favoriteRepository.delete(favorite);
-    }
-
-    public List<FavoriteResponse> findFavorites(Long memberId) {
-        List<Favorite> favorites = favoriteRepository.findAllByMemberId(memberId);
-
-        Map<Long, Station> stations = extractStations(favorites);
-
-        return favorites.stream()
-                .map(it -> toFavoriteResponse(stations, it))
-                .collect(Collectors.toList());
-    }
-
-    private FavoriteResponse toFavoriteResponse(Map<Long, Station> stations, Favorite favorite) {
-        return FavoriteResponse.of(
-                favorite,
-                StationResponse.of(stations.get(favorite.getSourceStationId())),
-                StationResponse.of(stations.get(favorite.getTargetStationId())));
-    }
-
-    private Map<Long, Station> extractStations(List<Favorite> favorites) {
-        Set<Long> stationIds = extractStationIds(favorites);
-        return stationRepository.findAllById(stationIds).stream()
-                .collect(Collectors.toMap(Station::getId, Function.identity()));
-    }
-
-    private Set<Long> extractStationIds(List<Favorite> favorites) {
-        Set<Long> stationIds = new HashSet<>();
-        for (Favorite favorite : favorites) {
-            stationIds.add(favorite.getSourceStationId());
-            stationIds.add(favorite.getTargetStationId());
-        }
-        return stationIds;
-    }
-
-}
