+++ src/test/java/nextstep/subway/favorite/acceptance/step/FavoriteAcceptanceStep.java
@@ -12,16 +12,17 @@ import java.util.Map;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class FavoriteAcceptanceStep {
-    public static ExtractableResponse<Response> 즐겨찾기_등록되어_있음(Long source, Long target) {
-        return 즐겨찾기_생성을_요청(source, target);
+    public static ExtractableResponse<Response> 즐겨찾기_등록되어_있음(Long source, Long target, String accessToken) {
+        return 즐겨찾기_생성을_요청(source, target, accessToken);
     }
 
-    public static ExtractableResponse<Response> 즐겨찾기_생성을_요청(Long source, Long target) {
+    public static ExtractableResponse<Response> 즐겨찾기_생성을_요청(Long source, Long target, String accessToken) {
         Map<String, String> params = new HashMap<>();
         params.put("source", source + "");
         params.put("target", target + "");
 
         return RestAssured.given().log().all().
+                auth().oauth2(accessToken).
                 contentType(MediaType.APPLICATION_JSON_VALUE).
                 body(params).
                 when().
@@ -31,8 +32,9 @@ public class FavoriteAcceptanceStep {
                 extract();
     }
 
-    public static ExtractableResponse<Response> 즐겨찾기_목록_조회_요청() {
+    public static ExtractableResponse<Response> 즐겨찾기_목록_조회_요청(String accessToken) {
         return RestAssured.given().log().all().
+                auth().oauth2(accessToken).
                 accept(MediaType.APPLICATION_JSON_VALUE).
                 when().
                 get("/favorites").
@@ -41,10 +43,11 @@ public class FavoriteAcceptanceStep {
                 extract();
     }
 
-    public static ExtractableResponse<Response> 즐겨찾기_삭제_요청(ExtractableResponse<Response> response) {
+    public static ExtractableResponse<Response> 즐겨찾기_삭제_요청(ExtractableResponse<Response> response, String accessToken) {
         String uri = response.header("Location");
 
         return RestAssured.given().log().all().
+                auth().oauth2(accessToken).
                 when().
                 delete(uri).
                 then().
